Clean up namespace of attrs (sunpy/sunpy-soar#184)
* Clean up namespace of attrs

* changelog
